- Fixed environment properties creation (adjusted enums and casting to enum)

+++ Engine/src/simulation/properties/property/api/PropertyType.java
@@ -1,5 +1,5 @@
 package simulation.properties.property.api;
 
 public enum PropertyType {
-    INT, DOUBLE, BOOLEAN, STRING
+    DECIMAL, FLOAT, BOOLEAN, STRING
 }
